updated gql endpoint one app.js
