+++ client/package.json
@@ -9,6 +9,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "antd": "^5.0.3",
     "graphql": "^16.6.0",
     "jwt-decode": "^2.2.0",
     "react": "^16.13.1",
